requirements.txt
python-telegram-bot==21.0
redis==5.0.1
python-dotenv==1.0.0
